refactor client to test util package
